Sidebar updates and a ton of renaming
