--- src/modals/FlightDetailsModal/FlightDetailsModal.definitions.tsx
@@ -1,6 +1,0 @@
-import type { FlightLegData } from "../../interfaces/flight-leg.interface";
-
-export interface FlightDetailsModalProps {
-  onClose: () => void;
-  flight: FlightLegData;
-}
