--- src/modals/PassengerDetailsModal/PassengerDetailsModal.definitions.ts
@@ -1,6 +1,0 @@
-import type { PassengerData } from "../../interfaces/passenger.interface";
-
-export interface PassengerDetailsModalProps {
-  passenger: PassengerData;
-  onClose: () => void;
-}
