+++ src/pages/PassengersPage/components/Passenger/Passenger.definitions.ts
@@ -1,0 +1,10 @@
+import type { PassengerData } from "../../../../interfaces/passenger.interface";
+
+/**
+ * Passenger Props
+ *
+ * @property {PassengerData} passenger - Passenger data
+ */
+export interface PassengerProps {
+  passenger: PassengerData;
+}
