+++ src/pages/PassengersPage/components/PassengerDetailsModal/PassengerDetailsModal.definitions.ts
@@ -1,0 +1,6 @@
+import type { PassengerData } from "../../../../interfaces/passenger.interface";
+
+export interface PassengerDetailsModalProps {
+  passenger: PassengerData;
+  onClose: () => void;
+}
