--- src/pages/RequestFlight/RequestFlight.tsx
@@ -1,14 +1,0 @@
-import FlightTimeSelector from "../../components/FlightTimeSelector/FlightTimeSelector";
-
-const RequestFlight = () => {
-  // Request flight tab
-  console.log("RequestFlight");
-
-  return (
-    <div>
-      <FlightTimeSelector />
-    </div>
-  );
-};
-
-export default RequestFlight;
