+++ src/pages/RequestFlightPage/RequestFlightPage.tsx
@@ -1,0 +1,14 @@
+import FlightTimeSelector from "./components/FlightTimeSelector/FlightTimeSelector";
+
+const RequestFlightPage = () => {
+  // Request flight tab
+  console.log("RequestFlight");
+
+  return (
+    <div>
+      <FlightTimeSelector />
+    </div>
+  );
+};
+
+export default RequestFlightPage;
