--- src/pages/SignIn/SignIn.tsx
@@ -1,8 +1,0 @@
-const SignIn = () => {
-  // Sign in page
-  console.log("SignIn");
-
-  return <div>SignIn</div>;
-};
-
-export default SignIn;
