+++ src/pages/SignInPage/SignInPage.tsx
@@ -1,0 +1,8 @@
+const SignInPage = () => {
+  // Sign in page
+  console.log("SignIn");
+
+  return <div>SignIn</div>;
+};
+
+export default SignInPage;
