--- src/pages/SignUp/SignUp.tsx
@@ -1,8 +1,0 @@
-const SignUp = () => {
-  // Sign in page
-  console.log("SignUp");
-
-  return <div>SignUp</div>;
-};
-
-export default SignUp;
