+++ src/pages/SignUpPage/SignUpPage.tsx
@@ -1,0 +1,8 @@
+const SignUpPage = () => {
+  // Sign in page
+  console.log("SignUp");
+
+  return <div>SignUp</div>;
+};
+
+export default SignUpPage;
