+++ src/pages/TripsPage/components/FlightDetailsModal/FlightDetailsModal.definitions.tsx
@@ -1,0 +1,6 @@
+import type { FlightLegData } from "../../../../interfaces/flight-leg.interface";
+
+export interface FlightDetailsModalProps {
+  onClose: () => void;
+  flight: FlightLegData;
+}
