+++ src/pages/TripsPage/components/FlightDetailsModal/FlightDetailsModal.module.css
