+++ src/pages/TripsPage/components/FlightDetailsModal/FlightDetailsModal.tsx
