+++ src/pages/TripsPage/components/FlightTicket/FlightTicket.definitions.tsx
@@ -1,0 +1,18 @@
+/* eslint-disable autofix/no-unused-vars */
+
+import type { FlightLegData } from "../../../../interfaces/flight-leg.interface";
+
+export interface FlightTicketProps {
+  flight: FlightLegData;
+}
+
+export enum LegType {
+  DEPARTURE = "Departure",
+  CONNECTING = "Connecting",
+  RETURN = "Return",
+}
+
+export enum FlightTicketColorVariant {
+  RED = "red",
+  BLUE = "blue",
+}
