+++ src/pages/TripsPage/components/FlightTicket/FlightTicket.module.css
